Split block action into two dedicated actions
